Update patch set 2
Patch Set 2:

> For the repro you will need EGL or GLES extension function that is
 > loaded by GetProcAddress() with number of function parameters
 > larger than number of available registers used in given calling
 > convention (thus it will fail more easily in i386 rather than
 > x86_64)
 > 
 > One of such extreme function is GetPerfCounterInfoINTEL from
 > INTEL_performance_query extension.

Right, but I think the request is for a failing test case to be added for Android. Can one be constructed so we can verify this behavior? As it stands, the graphics folks haven't weighed in, and I am not really qualified to write something to test this from scratch.

Patch-set: 2
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0
Label: Presubmit-Ready=0