Update patch set 1
Patch Set 1:

The 64-bit interface is defined in the library, while the underlying /dev/binder depends on the kernel.
When running the test, we push the test program to the device, and it is linked to the device's library. Therefore, the test don't fail unless it accesses /dev/binder directly. That's the reason I want to compile binderDriverInterfaceTest for 32-bit kernel.

Patch-set: 1
Reviewer: Gerrit User 1130827 <1130827@85c56323-6fa9-3386-8a01-6480fb634889>