Update patch set 7
Patch Set 7:

This is an alternative to https://android-review.googlesource.com/#/c/410645/ . The bp doesn't have duplicate targets, but the user has to specify an argument to test one of the interfaces.

The purpose of this CL is to fix the bug on oc-dev. It is not likely that we can add a requirement and ask the partners to immediately follow. Therefore I believe it is necessary to merge one of the CLs.

Patch-set: 7
Reviewer: Gerrit User 1130827 <1130827@85c56323-6fa9-3386-8a01-6480fb634889>