Update patch set 7
Patch Set 7:

I don't have a strong opinion on whether we support it or not, but I'd prefer not in order to simplify everyone's support. I'd prefer the other change (two binaries) if we're going to merge one of these changes.

We've been saying the 32-bit interface isn't supported on newer userspaces for a long time now: (but we haven't actually made that statement true)

https://android.googlesource.com/kernel/common/+/android-4.9/drivers/android/Kconfig#45

Patch-set: 7
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>