binder: free buffer as soon as possible
In current logic, we call sendReply() first, and then the Parcel buffer
is destructed where freeBuffer will be called. The BC_FREE_BUFFER cmd
will not be sent to kernel together with sendReply(). So there will
be a latency here.

I think it's better to free buffer asap, this change fixes this
by destructing Parcel buffer before sendReply() is called.

Change-Id: I0ef88864f4dbc63f3cdd2eba345b5dce6140b2d3
Signed-off-by: Ganesh Mahendran <opensource.ganesh@gmail.com>