binder: free buffer as soon as possible
In current code logic, we call sendReply() firstly, and then destruct
the Parcel buffer where freeBuffer will be called. So the BC_FREE_BUFFER
cmd will not be sent to kernel together with sendReply(). There will
be a latency here.

I think it's better to free buffer asap, this change fixes this
by destructing Parcel buffer before sendReply() is called.

Change-Id: I0ef88864f4dbc63f3cdd2eba345b5dce6140b2d3
Signed-off-by: Ganesh Mahendran <opensource.ganesh@gmail.com>