Update patch set 2
Patch Set 2:

(1 comment)

The problem here is that header-abi-dumper assumes that headers are all standalone / include what they need. I wish we lived in a world where that was always true :). But in this case it's not.

If header-abi-dumper can't be taught that certain headers depend on certain other headers being included first, then I guess we don't have a choice? My team will end up maintaining a delta from the upstream headers, if we can't get upstream to accept the change.

Patch-set: 2
Reviewer: Gerrit User 1010672 <1010672@85c56323-6fa9-3386-8a01-6480fb634889>