Update patch set 2
Patch Set 2:

Given Arve's concern, a systrace (perhaps to watch execution context switches, or to demonstrate the reduction in response latency) should to be inspected. If it does what you want (releases caller sooner) it could add a pair of context switches (WAG ~250ns on late model platforms) to the overall transaction processing. I expect it would have shown in the accumulated, albeit coarse, results; based on which sounds like the change is gratuitous. We really need to add a specific  servicemanager command end-to-end latency (send to reply) and transaction rate (send to buffer release) benchmark test in this CL given how subtle this is.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>