Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > I don't think we can expose more symbols than what we expose to the platform since these are stub libraries. At runtime, the libraries that are loaded are still the platform ones (just that anyone who compiles with the llndk only knows about the symbols that are exposed through the stub library).
> 
> Yup. But AFAIK we also don't expose symbols to the platform that we don't expose to llndk.
> 
> > So what we could do is, that we could check the platform library for abi breakages on all symbols (in exported headers, whether exposed to the llndk or not). Would that be okay ? And not too restrictive ? If yes, I can abandon this and take care of this by some hackery in soong, where I don't follow the usual llndk path for abi checks for these libraries.
> 
> Yes, that's fine. We don't expect to break ABI for these libraries. When you say "check for abi breakages", are you saying that for each symbol in the stub library, you check that the function prototype doesn't change? What about functions declared in the header that don't have corresponding symbols in the library? (Yet another GL weirdness -- headers declare all known extensions, but we don't bother providing symbols for many of them; apps query function pointers from the GL driver using eglGetProcAddress if the driver says it supports an extension.)

So how it works is that header-abi-dumper initially dumps all the function prototypes and corresponding type definitions of all parameters (structure layouts/ class layouts, vtables for C++ if applicable etc etc) seen in source files, filters out the ones in exported headers. Then, on a second pass we filter out the ones that don't appear in the .dynsym table of a library. So in this case those functions won't appear in the abi dump. It's okay to do anything to them. We check that any data type directly/ indirectly referenced by a function signature doesn't change. This will be a build breakage in the near future. Sounds fine ?

Patch-set: 2
Reviewer: Gerrit User 1121338 <1121338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0