Set default data location on non-FBE devices
After uninstall app's updates on non-FBE devices, change the settings in app of factory version
and then rebooting, the app setting is initialized.
The reason is that CE is set to default because neither directory is marked as default,
so the contents of DE path are deleted.
When neither CE nor DE is set as the default,
if the default is set according to flags, the app setting is retained.

Test: manual

Change-Id: I40800ad1609292ce05e85d5002622d2cf0533f0f
Signed-off-by: Hankyu Ji <hankyu.ji@lge.com>