Update patch set 1
Patch Set 1:

The alternative of verifying all interested apks is slow.  From my experiment on Pixel, verifying totally 100 MB of apks serially took 350ms.  100 MB was just I had at the moment, the actual size may be 200 MB or more, depending on what the vendors put in priv-apps.  It is harder to scale, and we'd just be postpone the problem to the future.

Patch-set: 1
Reviewer: Gerrit User 1042276 <1042276@85c56323-6fa9-3386-8a01-6480fb634889>