[installd] Test ClearAppData_InodeCache
Test purpose:
Installd stores an inode number of cache directory to xattr
of app data directory. When it cleans up the cache, the stored
inode number is used to solve the cache directory.

When app data is cleared, the stored inode number must be removed
from xattr. Otherwise it may delete a wrong directory which has
same inode number.

Bug: 69771310
Change-Id: I8827916697c96a253a04677d7895030b1918c2b6