Update patch set 2
Patch Set 2:

=== Finished presubmit run: L86900000125930787 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/540402?revision=2




NOTE: Since this build started, a new topic was set or a new patch was uploaded.

TreeHugger finished with: 23 passed.





WARNING: The tip-of-tree(4489854) for git_master/errorprone_fastbuild3e_linux may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4489854) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.












WARNING: The tip-of-tree(4489854) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4489854) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4489854) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4489854) for git_master/test_suites_x86_64_fastbuild3d_linux may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







Congratulations!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 2
Tag: autogenerated:TreeHugger:Verify-Stale-Pass