Accept EAGAIN on a binder transaction.
A recent change to the binder kernel driver delays the
delivery of BR_TRANSACTION_COMPLETE until a reply comes
in. The binderDriverInterfaceTest uses the non-blocking
interface, and assumes BR_TRANSACTION_COMPLETE comes in
on the first read. I don't think we depend on this
assumption in any of our real code, so change the test.

Bug: 69442864
Test: binderDriverInterfaceTest passes
Change-Id: Iafca061c1bb7badd6cb61b3e876d9c3bbfa18542