Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Lloyd/Dom, would you mind taking a look at this to determine whether it fits in with your plans? I'm fine taking it if it's moving the right direction, but I don't really want to complicate things unnecessarily.

I think the patch won't bother us much.

Longer term, it would be cleaner and more scalable for use on other displays if instead of a property we obtained the adjustment from the hardware composer interface. While it could be added as a mode attribute, I expect it would be better done as a new function that queries mode-independent display attributes. We could certainly get the display EDID through the same function with a different attribute selector.

I expect we could take this patch as is for now, and when we have the new API, require it be reported through the API when hardware composer implementation reports it supports the new API. We could still read the property ourselves when supporting an older implementation (though I'm not sure how we handle an EDID well with an older implementation)

Patch-set: 1
Reviewer: Gerrit User 1086703 <1086703@85c56323-6fa9-3386-8a01-6480fb634889>