Update patch set 1
Patch Set 1:

No. Just checking to make sure it's not stale.

I'm actually thinking, for things built with the VNDK, that after this, we should automatically select vndbinder or binder based on if it is a vendor binary or not.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0