Add missing sensor types to <android/sensor.h>.
This matches Sensor.java
(https://developer.android.com/reference/android/hardware/Sensor.html).

I've included the types for a few of the easy cases, but I've added a
link to the Java documentation for the rest: the Java documentation is
much better and actively maintained. Duplicating it here doesn't seem
worthwhile, and <android/multinetwork.h> already takes this approach.

Bug: https://github.com/android-ndk/ndk/issues/621
Test: builds
Change-Id: Icffd85071e56bd7f45e8f3dc9d300a495859dbcb