DL policy extension for SchedulingPolicyService [DO NOT MERGE]
Enabling of SCHED_DEADLINE policy through the SchedulingPolicyService by
exporting the requestPriorityDL.
This function takes as input the runtime, deadline and period parameters
that will be used to schedule the process with the given tid, instead of
just the priority parameter used for FIFO/RR scheduling.

This function performs an initial parameters checking so that

1024 <= runtime <= deadline <= period

Here, 1024 is a magic number defined by the kernel, while the other
constraints are necessary for the schedulability of the task.

Test: manual - experimental extension used for prototyping
Change-Id: I71d40c411cdddd1ed0e6b1b8d94049f3306e7ca7