Allow a primary display disconnect (HWC1)
This is the HWC1 implementation of patch made to the HWC2 code path with
a similar description.

This patch forwards the primary display disconnect event to the
Framework, and otherwise ensures that SurfaceFlinger does not crash
while there is no primary display.

Note however that the Framework code may not respond correctly to this
event.

Bug: 38464421
Test: Builds (Untested!)

Change-Id: I68766b69ed4ad5bf08c19c931c290b7cfa68c7bd
(cherry picked from commit 2ae2b3bd5bfbacc3b52f222441bfbbeb9bd54dc7)