Move hotplug processing to the main thread (HWC1)
This is the HWC1 implementation of patch made to the HWC2 code path with
a similar description.

Queue up all hotplug events for processing in the main thread, as part
of a display transaction (eDisplayTransactionNeeded).

This is needed so that everything done for each individual hotplug disconnect or
connect is done at the same time,  such as creating and destroying the
corresponding DisplayDevice.

This fixes an issue with a hotplug disconnect event followed by an
immediate connect event for the same display not being handled properly.

Bug: 38464421
Test: Builds (Untested!)
Change-Id: I8e43a0a96c68b991dc8df3b3e4390e5fdee24f20
(cherry picked from commit ba04e6237fb6ae1aede76b543028da101412b11d)