Update patch set 1
Patch Set 1:

So a couple of notes that I wanted to address:

Using only this mechanism, touchpad events will be sent via onGenericMotionEvent rather than onCapturedPointerEvent.  We could have it sent via onCapturedPointerEvent if we added filtering in ViewRootImpl for a touchpad source with pointer capture.

Another problem we are running into currently is that the touchpad events are not causing a view to regain focus, which means that after leaving the app and coming back all touchpad events now go to the activity rather than the view.  I'm currently looking into this, but I wanted to get preliminary feedback on the approach.

Patch-set: 1
Reviewer: Gerrit User 1156119 <1156119@85c56323-6fa9-3386-8a01-6480fb634889>