Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> So a couple of notes that I wanted to address:
> 
> Using only this mechanism, touchpad events will be sent via onGenericMotionEvent rather than onCapturedPointerEvent.  We could have it sent via onCapturedPointerEvent if we added filtering in ViewRootImpl for a touchpad source with pointer capture.
> 
> Another problem we are running into currently is that the touchpad events are not causing a view to regain focus, which means that after leaving the app and coming back all touchpad events now go to the activity rather than the view.  I'm currently looking into this, but I wanted to get preliminary feedback on the approach.

So a bit of an update after some more digging.

The root of this comes down to view focus.  Events delivered over onGenericMotionEvent and onCapturedPointerEvent  are not sent when the view doesn't have focus.  By default, it would appear that views are not focusable while in touch mode.  This isn't a problem for captured pointers as mice seem to implement a virtual d-pad, which triggers a KeyEvent, which puts the device in non-touch mode, which causes focusable (just not in touch mode) views to come into focus.  The primary joystick on a gamepad does the same thing (virtual d-pad).

Our touchpad change doesn't cause any KeyEvents to be generated by the touchpad.  So using the touchpad while captured doesn't cause the device to leave touch mode, which doesn't allow a view to come into focus.

This whole thing is can be avoided by allowing your OpenGL or Vulkan view to be focusable in touch mode, but that doesn't seem to be a great answer.  If I haven't misunderstood the mechanism by which the pointer capture is causing the device to leave touch mode, it seems somewhat convoluted.

I feel like the best way to address this would be to cause the device to leave touch mode on generic events and captured pointer events, but I don't know what other kinds of input other than touchpads and gamepads would come over generic motion events.  Looking for any feedback!

Patch-set: 1
Reviewer: Gerrit User 1156119 <1156119@85c56323-6fa9-3386-8a01-6480fb634889>