Update patch set 1
Patch Set 1:

From the commit message, it seems #1 is unused only because you always cancel it.  With or without this change, BufferQueue always allocates #1, #2, and #3.  Is waste of memory a real concern?

That said, if you look at (8), it is possible that a consumer fence is still associated with #0 and the rendering to #0 requires a fence wait.  If the canceled #1 was returned to the head of queue in (7), #1 would be dequeued in (8) and the fence wait would be avoided.  Is this what you are seeing?

Patch-set: 1
Reviewer: Gerrit User 1096171 <1096171@85c56323-6fa9-3386-8a01-6480fb634889>