Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > No fence in the display driver?  That means everything is
 > synchronous to SurfaceFlinger!
 > >
 > > For example, when SurfaceFlinger calls presentDisplay to replace
 > #0 by #1 as the scan-out buffer, it assumes the present takes
 > effect immediately, and #0 is ready for reuse.  That is apparently
 > not the case since #0 is not replaced by #1 until the next vsync.
 > >
 > > That would explain the screen tearing because the producer may
 > start rendering to #0 before the next vsync.  That might also
 > explain why you want the producer to use the previously canceled
 > buffer over #0.
 > >
 > > This all sounds dangerous to me..
 > 
 > Actually, the standard FB/KMS(less than linux 4.9) doesn't support
 > fence. And three framebuffers are used to avoid tearing issue. That
 > is traditional way.
 > I think keep backward compatibility is also important.

I'm not on the Android graphics team, so I'm not speaking authoritatively here, and I don't have any deep knowledge about the implementation of hardware composition and rendering. That said:

While this small tweak might help your implementation work today, most every part of the Android graphics buffer infrastructure strongly assumes that the consumer (the FB in this case) only accesses buffers it has acquired and not yet released, and that it respects fences when they are provided.  

Doing anything else is extremely fragile, since behavior and code changes will be made that assume the above is true.

If you need to run on top of FB/KMS, then you likely need to write a wrapper layer around them that ensures fences are waited on and that buffers that are released are no longer going to be accessed.  Nothing else will be safe.

Patch-set: 1
Reviewer: Gerrit User 1010050 <1010050@85c56323-6fa9-3386-8a01-6480fb634889>