vulkan: Allow AHB extension to get to the driver
We don't need to do anything special with the AHB extension at the
driver layer, but it's defined as a "known extension" so that we can
export symbols for it. So it didn't fall into the generic extension
path, but we also didn't add it to the list to pass on to the driver.
So drivers were never seeing that the extension was enabled.

Remove switch default cases, so that -Wswitch can catch bugs like this
in the future.

Test: in-progress MediaVulkanGpuTest (ag/3789901/18)
Bug: 77297007
Change-Id: I3fe460240cfd1f868b7275abbb0289b806bc7603
(cherry picked from commit 19cc7541a88e2849e1efb8966e0356d15a965db8)