submit patch to google for TCL CTS When we run our handset with CTS verifier8.1_R2 based on Android8.1.0 on MSM8996,we met 1 failed case.This case cannot be tested in China product. Test case: 1. Device Owner Test -> Policy transparency test -> Disallow config cell broadcast
To test this case, the option "Emergency broadcasts" should shows in  "Settings->Sound".
Since this is a product for China market and the emergency broadcast is not supported in China, so we removed this option "Emergency broadcasts" from settings. Then the test failed.

And Andyds from Google ask us for patch. So we submit patches.