Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > It seems like this should also be conditional on whether a looper thread has actually been created. Now it will be wrong for processes without a looper thread. However, maybe we should just consider a looper thread as not executing (current behavior) because there is no guarantee that it will start processing binder transactions anytime soon. We can't be sure, so I think it's better to err on the side of printing longer starvation times. Anyone else have thoughts on this?
> 
> thanks for your reply, I have updated detailed description of the binder thread pool.
> the core of the bug is 1st binder thread not counted in mMaxThreads.

I think we should fix this by making mMaxThreads what it actually is; if mMaxThreads is 15, there should be only 15 binder threads, not more. We have a similar fix in libhwbinder, where we tell the kernel to spawn one less thread than what was requested.

So that fix would change DEFAULT_MAX_BINDER_THREADS to 16, and change the argument of the ioctl to use mMaxThreads - 1 (since userspace already spawns one - either by calling startThreadPool() or by using a poll thread).

My main worry with that change is that there may be out of tree code calling setThreadPoolMaxThreadCount(0) (to effectively get a threadpool size of 1), and we'd be breaking those. Perhaps if the arg is zero, we can treat it as 1.

Patch-set: 3
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>