Update patch set 3
Patch Set 3:

(6 comments)

PTAL

In this patch set, int in ParcelFileDescriptor's ctor, reset() and release()  are replaced with unique_fd.
Please note that this interface may make a subtle pitfall when releasing the ownership of a file descriptor because of unique_fd's implicit operator int:
  ParcelFileDescriptor pfd;
  ...
  int fd = pfd.release(); // This should be pfd.release().release()!
  DoSomethingAndCloseFd(fd); // fd is already invalid.

I'm OK with either using int or unique_fd in the interface, which do you prefer? WDYT?

> Patch Set 1:
> 
> It's already only one call to read and write file descriptors that can be handled as ParcelFileDescriptors in Java. What value is this adding?

The main difference between ParcelFileDescriptor and FileDescriptor is that, when returning a file descriptor, Java code generated from AIDL closes the original file descriptor for the former[1] while not doing it for the latter[2].
In other words, using FileDescriptor instead of ParcelFileDescriptor as a return type in AIDL can result in leaking file descriptors.
This is especially problematic when dealing with pipes or when using StorageManager.openProxyFileDescriptor because not closing a FD can cause deadlock in the process which holds the read end.

[1] https://android.googlesource.com/platform/frameworks/base/+/009b8521338b6c89335563c961ec9beefec9b5c7/core/java/android/os/ParcelFileDescriptor.java#1010
[2] https://android.googlesource.com/platform/frameworks/base/+/60f55a255f1cd379bfc6cbc9122c43d48925f7c4/core/java/android/os/Parcel.java#723

> 
> I'd much rather see versions of read/writeParcelFileDescriptor that use unique_fd than add yet another file descriptor wrapper to the C++ side.

Added writeDupParcelFileDescriptor() and readUniqueParcelFileDescriptor to Parcel and rewrote ParcelFileDescriptor with them.
Also added tests for the new Parcel methods.

Patch-set: 3
Reviewer: Gerrit User 1069219 <1069219@85c56323-6fa9-3386-8a01-6480fb634889>