Update patch set 1
Patch Set 1:

(2 comments)

Jiyong, could you extend this to Java and flesh out a user of the API with b/110759835? I think if we had more information about what we were trying to do here, we might be able to understand why this is needed more. As it stands, it doesn't seem very useful yet.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>