Bugreport: Adding eBPF maps content to bugreport
The eBPF maps created by netd trafficController are used to store the
detail network traffic stats since device boot. It was not included in
the netd default dumping content since the detail content of the maps
would flood the basic information dumped and may not be useful in
general cases. But for bug report we and detail dumpstate operation we
should include them to help track down networking related issues.

Bug: 112226716
Test: dumpstate show map content, bug report have the section showing
content of each eBPF maps.

Change-Id: Id90fdbf67a43e5e2d8c76e65d00976a48b511a33