Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > Could you give a concrete example of a case this will fix?
> > > > 
> > > > We do not resolve the realpaths by intention. That's because there are cases when this will not work properly.
> > > 
> > > Hi Juravle:
> > > In our case, the preinstalled apps are placed under system/partner-app. When phone bootup, the soft links from data/app to system/partner-app will be set up.
> > > 
> > > Can u tell me the cases that will not work properly?
> > 
> > When loading oat files the runtime does not resolve the real paths and as such you may end up in fallback mode (extracting the apk in memory) depending on the use case.
> 
> I can't think of such an example, could u give a concrete example?

When loading the file the runtime checks the location of the dex file specified during loading, with the location of the dex files used during compilation. If they don't match the oat file will no be loaded. As such, if the compiled with a real path but then loaded with a sym link things might work as you expect.

More, there are corner case, like files being on a system partition which is not writable, files which still needs to be optimized. Trying to compile against a realpath will fail for those as the expected location is not writable.

Patch-set: 3
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>