Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > > Patch Set 3:
> > > > > 
> > > > > Could you give a concrete example of a case this will fix?
> > > > > 
> > > > > We do not resolve the realpaths by intention. That's because there are cases when this will not work properly.
> > > > 
> > > > Hi Juravle:
> > > > In our case, the preinstalled apps are placed under system/partner-app. When phone bootup, the soft links from data/app to system/partner-app will be set up.
> > > > 
> > > > Can u tell me the cases that will not work properly?
> > > 
> > > When loading oat files the runtime does not resolve the real paths and as such you may end up in fallback mode (extracting the apk in memory) depending on the use case.
> > 
> > I can't think of such an example, could u give a concrete example?
> 
> When loading the file the runtime checks the location of the dex file specified during loading, with the location of the dex files used during compilation. If they don't match the oat file will no be loaded. As such, if the compiled with a real path but then loaded with a sym link things might work as you expect.
> 
> More, there are corner case, like files being on a system partition which is not writable, files which still needs to be optimized. Trying to compile against a realpath will fail for those as the expected location is not writable.

I think the root cause is that PMS doesn't get the real path, but runtime gets the real path.
in art\runtime\oat_file_assistant.cc file,line is 93,
  UniqueCPtr<const char[]> dex_location_real(realpath(dex_location, nullptr));
dex_location_real will get real path.

PMS and runtime should have the same behavior to get the location of dex file.

Supposed that a app is at path A of system.img, the dex files of the app during compilation is at path A/oat. 
When phone bootup, the soft links from path A to path B will be set up.
The real location of the app dex files are at A,not B. If not to get the real path,the path is not what we expect. 

Another example,if the apps are placed under system/app, the path of system/app is read-only, so the location of the dex files will be at data/dalvik-cache,not at system/app.
The apps can be optimized correctly.

Patch-set: 3
Reviewer: Gerrit User 1127267 <1127267@85c56323-6fa9-3386-8a01-6480fb634889>