Update patch set 5
Patch Set 5:

Hi Tang,

Thank you for the patch. Could you please elaborate a bit more on the actual bug? Is there a tracking bug for this on the issuetracker? Can you add tests to tests/InputDispatcher_test.cpp for this case? What is the easiest way to reproduce this problem?

Also, please format the commit message with a title and description.

Patch-set: 5
Reviewer: Gerrit User 1149384 <1149384@85c56323-6fa9-3386-8a01-6480fb634889>