libbinder_ndk: Add C++ wrapper code.
This library wraps libbinder_ndk with types similar to those found in
libbinder. This makes implementations more familiar and reduces the
amount of auto-generated code that has to be repeated in AIDL.

In order to simplify the distribution of the this code, it is
header-only in the NDK headers. This allows NDK users to use AIDL
against the NDK with no additional dependencies.

Bug: 112664205
Test: wip

Change-Id: I6e6ef646226d474d538639f740e340c4e37fa7f7