Update patch set 23
Patch Set 23:

(2 comments)

> Is it supposed to go in somewhere other than the header? I added documentation to AParcel.h. AIBinder.h was already documented (and I have added more according to your comments).

in the header is correct, AParcel just happened to be the first one I checked when I then made that comment. There's still some methods missing documentation in a few of these. This is all a developer gets, so lifecycle needs to be explicitly documented as well as any error return values.

Patch-set: 23
Reviewer: Gerrit User 1010103 <1010103@85c56323-6fa9-3386-8a01-6480fb634889>