Update patch set 29
Patch Set 29: Verified+1 Presubmit-Ready+1

If you guys are okay with this, I'd like to do incremental development on this in order to prevent having a CL like this again. I've broken out the pieces adding this to the NDK. I'm happy to break this apart as well into several smaller CLs if that would help as well.

Patch-set: 29
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1