Update patch set 1
Patch Set 1:

(2 comments)

I had discussed this with Jiyong, and we agreed it was a bit much given the semantics of these methods. I can add it (and will do so universally) if you desire so, but I'd like the entire surface to be consistent.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>