libbinder: BBinder/BpBinder detachObject cleans up object.
IBinder has an API on it which attaches/detaches objects. This API is
given a key, value, and cleanup function. When an IBinder is destructed,
the object is also destroyed. However, detachObject just doesn't
actually delete the associated object (rather it leaks the object).

Both BBinder and BpBinder use BpBinder::ObjectManager to implement this
API.

A releaseObject API would make a lot of sense. In fact, if detachObject
returned a void*, it would be the preferred solution to just add something
like __attribute__((warn_unused_result)) here to embody the ownership
model and cause the client to clean it up. As it stands, BpBinder/BBinder
guard accesses to this data with their respective locks. These locks are
not available to the client. For this reason, we cannot expect a client to
findObject and then detachObject since it isn't really guaranteed to be
the same object.

Bug: 114127243
Test: boot Pixel 2 to screen and inspect logs (this API is hardly used
    which is why this was hardly noticed).
Test: manually use this instead of the workaround in libbinder_ndk
    which was instated at the discovery of this bug.

Change-Id: I86893d1ac60ece5f41025b3e47b5662e0eb6418a