surfaceflnger: Avoid stale content in framebuffer
When the layer list given to hwcomposer is empty, hwcomposer is expected
to simply present the contents of the framebuffer target buffer. So in
this scenario, surfaceflinger needs to ensure the framebuffer target
content is not stale prior to hand-off to hwcomposer.

Previously the HWC2 paths in surfaceflinger failed to do this, resulting
in display of stale content during certain UI transitions. For example,
during app startup, the layer list may briefly be empty; as the
framebuffer target was not cleared, content from some older frame was
shown on display.

Bug 200428932

Change-Id: I09951bdcd28c270f2d32650b63e8c23d3523c08a