Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> All of these APIs are new in Q, and I haven't added those guards in other files. Is it still necessary?

Yes, because unlike the SDK the NDK just has a single version regardless of what is being targeted. So the headers need to scope themselves to what the user is building against.

> If it is, I'd like to do it in all of these files in a separate CL.

Ack, but mind handling this file in this CL since it's new?

Patch-set: 5
Reviewer: Gerrit User 1010103 <1010103@85c56323-6fa9-3386-8a01-6480fb634889>