sf: Fix issue with dequeueBuffer when multiple display present
swapBuffers is called even for device composition on MDP3 target
based on flip client target request. dequeueBuffer call is skipped
as there is no client composition. This is fine for single display
but if multiple display i.e. vds is present, this will be an issue
as current surface will not be switched and swapBuffers will fail.

When flip client target request is present, make sure buffer is
dequeued before swapBuffers.

CRs-Fixed: 2311294

Change-Id: I933cbae2f09f81eef6555b1bb1e5991d2c450930