Add an abstract attribute for BufferLayer
Some surfaces do not need to allocate buffer, because they are
abstract layers. They use to propagate transform, crop and so on.
They may have large width and height. These size of width and height
may be over the texture limitation of GPU. SurfaceFlinger would
block creating this layer. They should not be limited by capability
of GPU. Therefore, we add a creation flag to ignore the check.

Change-Id: I00a33a0420fbc19f4664106ae00eed1fbda95421